pyproject.toml
[tool.poetry]
name = "picturebot"
version = "0.1.0"
description = ""
authors = ["Replit user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
python-dotenv = "^0.15.0"
replit = "^2.0.0"
timedelta = "^2020.12.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
